+++ packages/firebase_transport/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* chore: improve test coverage.
+
 ## 1.0.0
 
 * Initial release.