+++ packages/notification_transport/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 * chore: update example pubspec.lock.
+* chore: improve test coverage.
 
 ## 1.0.0